# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  version: "1.0.0"
  description: API for WASAText messaging application - A comprehensive chat system with user authentication, conversations, messages, and group management features
  contact:
    name: WASA Team
    email: wasa@uniroma1.it
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
...
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - bearerAuth: []
paths:
  /session:
    post:
      tags: ["authentication"]
      summary: User login and session creation
      description: Authenticates a user and returns an identifier token. If the user does not exist, it will be created automatically with the provided name.
      operationId: doLogin
      security: []
      requestBody:
        description: User login credentials
        required: true
        content:
          application/json:
            schema:
              type: object
              description: User login information
              properties:
                name:
                  type: string
                  description: Username for login
                  example: "MariaUser123"
                  pattern: '^[a-zA-Z0-9_]{3,16}$'
                  minLength: 3
                  maxLength: 16
              required:
                - name
      responses:
        '201':
          description: User successfully logged in
          content:
            application/json:
              schema:
                type: object
                description: Login response containing user identifier
                properties:
                  identifier:
                    type: string
                    description: Unique identifier token for the user
                    example: "abcdef0123456789"
                    pattern: '^[a-zA-Z0-9]{12,32}$'
                    minLength: 1
                    maxLength: 32
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/username:
    put:
      tags: ["users"]
      summary: Update user username
      description: Updates the current authenticated user's username to a new value.
      operationId: setMyUserName
      requestBody:
        description: New username information
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Username update information
              properties:
                username:
                  type: string
                  description: New username for the user
                  example: "newusername123"
                  pattern: '^[a-zA-Z0-9_]{3,16}$'
                  minLength: 3
                  maxLength: 16
              required:
                - username
      responses:
        '200':
          description: Username updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  message:
                    type: string
                    description: Success message
                    example: "Username updated successfully"
                    minLength: 1
                    maxLength: 200
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/photo:
    put:
      tags: ["users"]
      summary: Update user profile photo
      description: Updates the current authenticated user's profile photo by uploading an image file.
      operationId: setMyPhoto
      requestBody:
        description: Photo file upload
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: Photo upload information
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file (JPEG, PNG, GIF) for user profile photo
              required:
                - file
      responses:
        '200':
          description: Photo updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Success response with photo URL
                properties:
                  message:
                    type: string
                    description: Success message
                    example: "Photo updated successfully"
                    minLength: 1
                    maxLength: 200
                  photoUrl:
                    type: string
                    description: URL of the uploaded photo
                    example: "/photos/user_abc123_def456.jpg"
                    pattern: '^/photos/user_[a-zA-Z0-9_-]+\\.(jpg|jpeg|png|gif)$'
                    minLength: 15
                    maxLength: 100
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /conversations:
    get:
      tags: ["conversations"]
      summary: Get user conversations
      description: Retrieves all conversations for the current authenticated user including both direct messages and group conversations.
      operationId: getMyConversations
      responses:
        '200':
          description: List of conversations retrieved successfully
          content:
            application/json:
              schema:
                type: array
                description: Array of user conversations
                minItems: 0
                maxItems: 100
                items:
                  $ref: '#/components/schemas/Conversation'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: ["conversations"]
      summary: Create new conversation
      description: Creates a new conversation (either direct message or group conversation) for the authenticated user.
      operationId: createConversation
      requestBody:
        description: Conversation creation data
        required: true
        content:
          application/json:
            schema:
              type: object
              description: New conversation information
              properties:
                name:
                  type: string
                  description: Name of the conversation
                  example: "New Chat Group Discussion"
                  minLength: 1
                  maxLength: 100
                type:
                  type: string
                  description: Type of conversation
                  example: "group"
                  enum: ["direct", "group"]
                  minLength: 4
                  maxLength: 6
                  default: "group"
              required:
                - name
      responses:
        '201':
          description: Conversation created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDetail'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /conversations/{id}:
    get:
      tags: ["conversations"]
      summary: Get conversation details
      description: Retrieves detailed information about a specific conversation including all messages and comments.
      operationId: getConversation
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the conversation
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_-]{8,32}$'
            minLength: 8
            maxLength: 32
            example: "conv_abc123def456"
      responses:
        '200':
          description: Conversation details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDetail'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /messages:
    post:
      tags: ["messages"]
      summary: Send message to conversation
      description: Sends a new message to a specified conversation for the authenticated user.
      operationId: sendMessage
      requestBody:
        description: Message data
        required: true
        content:
          application/json:
            schema:
              type: object
              description: New message information
              properties:
                conversationId:
                  type: string
                  description: Target conversation identifier
                  example: "conv_abc123def456"
                  pattern: '^[a-zA-Z0-9_-]{8,32}$'
                  minLength: 8
                  maxLength: 32
                body:
                  type: string
                  description: Message content
                  example: "Hello everyone!"
                  minLength: 1
                  maxLength: 2000
                  pattern: '^.+$'
              required:
                - conversationId
                - body
      responses:
        '201':
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /messages/{id}/forward:
    post:
      tags: ["messages"]
      summary: Forward message to another conversation
      description: Forwards an existing message to another conversation for the authenticated user.
      operationId: forwardMessage
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the message to forward
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_-]{8,32}$'
            minLength: 8
            maxLength: 32
            example: "msg_abc123def456"
      requestBody:
        description: Forward destination information
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Forward destination information
              properties:
                conversationId:
                  type: string
                  description: Target conversation identifier
                  example: "conv_def456ghi789"
                  pattern: '^[a-zA-Z0-9_-]{8,32}$'
                  minLength: 8
                  maxLength: 32
              required:
                - conversationId
      responses:
        '200':
          description: Message forwarded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Message or conversation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /messages/{id}/comment:
    post:
      tags: ["messages"]
      summary: Add comment to message
      description: Adds a comment to an existing message for the authenticated user.
      operationId: commentMessage
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the message to comment on
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_-]{8,32}$'
            minLength: 8
            maxLength: 32
            example: "msg_abc123def456"
      requestBody:
        description: Comment data
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Comment information
              properties:
                comment:
                  type: string
                  description: Comment content
                  example: "Great point! I totally agree with this message!"
                  minLength: 1
                  maxLength: 500
                  pattern: '^.+$'
              required:
                - comment
      responses:
        '200':
          description: Comment added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /messages/{id}/uncomment:
    post:
      tags: ["messages"]
      summary: Remove comment from message
      description: Removes a comment from a message for the authenticated user.
      operationId: uncommentMessage
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the message to remove comment from
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_-]{8,32}$'
            minLength: 8
            maxLength: 32
            example: "msg_abc123def456"
      responses:
        '200':
          description: Comment removed successfully
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  message:
                    type: string
                    description: Success message
                    example: "Comment removed successfully"
                    minLength: 1
                    maxLength: 200
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Message or comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /messages/{id}:
    delete:
      tags: ["messages"]
      summary: Delete message
      description: Deletes a message for the authenticated user. Only the message author can delete their own messages.
      operationId: deleteMessage
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the message to delete
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_-]{8,32}$'
            minLength: 8
            maxLength: 32
            example: "msg_abc123def456"
      responses:
        '204':
          description: Message deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication. Include this token in the Authorization header as 'Bearer {token}'
  schemas:
    Error:
      type: object
      description: Standard error response format
      properties:
        error:
          type: string
          description: Error message describing what went wrong
          example: "An error occurred"
          minLength: 1
          maxLength: 500
      required:
        - error
    Conversation:
      type: object
      description: Basic conversation information
      properties:
        id:
          type: string
          description: Unique conversation identifier
          example: "conv_abc123def456"
          pattern: '^[a-zA-Z0-9_-]{8,32}$'
          minLength: 8
          maxLength: 32
        name:
          type: string
          description: Conversation name
          example: "My Chat Room Discussion"
          minLength: 1
          maxLength: 100
        type:
          type: string
          description: Conversation type
          example: "group"
          enum: ["direct", "group"]
          minLength: 4
          maxLength: 6
          pattern: '^(direct|group)$'
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2023-01-01T12:00:00Z"
      required:
        - id
        - name
        - type
    ConversationDetail:
      type: object
      description: Detailed conversation information including messages
      properties:
        id:
          type: string
          description: Unique conversation identifier
          example: "conv_abc123def456"
          pattern: '^[a-zA-Z0-9_-]{8,32}$'
          minLength: 8
          maxLength: 32
        name:
          type: string
          description: Conversation name
          example: "My Chat Room Discussion"
          minLength: 1
          maxLength: 100
        type:
          type: string
          description: Conversation type
          example: "group"
          enum: ["direct", "group"]
          minLength: 4
          maxLength: 6
          pattern: '^(direct|group)$'
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2023-01-01T12:00:00Z"
        messages:
          type: array
          description: Array of messages in the conversation
          minItems: 0
          maxItems: 1000
          items:
            $ref: '#/components/schemas/Message'
      required:
        - id
        - name
        - type
        - messages
...